refactor(orchestrator-ui): extract useOrchestratorData live-data hook from the workbench (#9960)
Continues the OrchestratorWorkbench decomposition the issue asks for ("decompose
into a live-data wrapper + presentational sub-views"). All task/session DATA
state (status, tasks, the selected task's detail + timeline) and the
fetch / poll / SSE-stream / mutation logic move into a new useOrchestratorData
hook; the component keeps the UI state (selection, filters, drawers) and feeds
it in. Behavior preserved exactly (same fetchers, polling cadence, stream
debounce, cross-task guards, runMutation reconcile).

- new use-orchestrator-data.ts (hook + the data-only helpers it owns:
  mergeById / splitTimelineItems / backend-absent + error classifiers + poll
  constants), so the data flow is testable without rendering the view.
- new use-orchestrator-data.test.ts (5): list+status on mount, backendAbsent on
  404, detail+timeline on selection, clear-on-deselect, loud actionError on
  mutation failure — the wrapper test the stateful layer never had.
- OrchestratorWorkbench.tsx: 3,931 → ~3,440 lines across this + the earlier
  glyph extraction; export contract unchanged.

Verified: typecheck clean (only the pre-existing FilterSelect ref error),
view bundle builds, inspector + view + panel suites green (23), hook test green
(5).
